Show some Love, community!
Python Program to Calculate Which Roles a Business Needs to Grow and How Many FTEs are in a Particular Role. To Calculate "Customer Lifetime Value." To Test Logic Over Time. Ultimately, to Optimise a Business. This information is relevant to Business Acquisitions, New Business Owners, and Business Owners Looking to Expand.